tests/wpt/web-platform-tests/css/cssom-view/MediaQueryListEvent.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta name="flags" content="dom">
<title>CSS Test: CSSOM View MediaQueryListEvent</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://drafts.csswg.org/cssom-view/#mediaquerylistevent">
<div id="log"></div>
<script>
"use strict";

test(t => {
    assert_equals(new MediaQueryListEvent("test").type, "test");
}, 'type can be different from "change"');

test(t => {
    const event = new MediaQueryListEvent("change");

    assert_equals(event.media, "");
    assert_false(event.matches);
    assert_false(event.bubbles);
    assert_false(event.cancelable);
}, "init dictionary default values");

test(t => {
    const event = new MediaQueryListEvent("change", {
        media: "test",
        matches: true,
        bubbles: true,
        cancelable: true,
    });

    assert_equals(event.media, "test");
    assert_true(event.matches);
    assert_true(event.bubbles);
    assert_true(event.cancelable);
}, "init dictionary overrides");
</script>
